dynamic width

+++ gum/indexer.py
@@ -178,13 +178,13 @@ class Indexer(object):
                 mapping_type.index_document(instance)
                 if stdout:
                     import os
-                    progress = int((step + 1) / float(total_instances))
+                    progress = (step + 1) / float(total_instances)
                     _, columns = os.popen('stty size', 'r').read().split()
                     limit = min(int(columns) - 10, 100)
-                    graph_progress = progress * limit
+                    graph_progress = int(progress * limit)
                     stdout.write('\r', ending='')
                     progress_format = "[%-{}s] %d%%".format(limit)
-                    stdout.write(progress_format % ('='*graph_progress, progress*100), ending='')
+                    stdout.write(progress_format % ('='*graph_progress, int(progress*100)), ending='')
                     stdout.flush()
             if stdout:
                 stdout.write('')
